+++ client/src/components/Layout/index.jsx
@@ -1,9 +1,9 @@
 import React, { useContext } from 'react';
-import { Link } from 'react-router-dom';
 
 import { faBook } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Header, Issue } from './styled';
+import { CLIENT_URL } from '../../config';
 
 import { UserContext } from '../../stores/userStore';
 
@@ -15,12 +15,15 @@ const Layout = () => {
   if (!name) {
     return null;
   }
+
+  const clickHandler = () => {
+    window.location.href = CLIENT_URL;
+  };
+
   return (
     <Header>
       <FontAwesomeIcon icon={faBook} size="2x" color="white" />
-      <Link to="/">
-        <Issue>Issues</Issue>
-      </Link>
+      <Issue onClick={() => clickHandler(true)}>Issues</Issue>
     </Header>
   );
 };
