+++ client/src/components/Sidebar/Label/styled.js
